GLTF: Organize structures into a subfolder
