+++ modules/gltf/SCsub
@@ -7,5 +7,7 @@ env_gltf = env_modules.Clone()
 
 # Godot's own source files
 env_gltf.add_source_files(env.modules_sources, "*.cpp")
+env_gltf.add_source_files(env.modules_sources, "extensions/*.cpp")
+env_gltf.add_source_files(env.modules_sources, "structures/*.cpp")
 if env["tools"]:
     env_gltf.add_source_files(env.modules_sources, "editor/*.cpp")
