+++ modules/gltf/extensions/gltf_spec_gloss.cpp
