+++ modules/gltf/gltf_document_extension.h
@@ -31,8 +31,8 @@
 #ifndef GLTF_DOCUMENT_EXTENSION_H
 #define GLTF_DOCUMENT_EXTENSION_H
 
-#include "gltf_node.h"
 #include "gltf_state.h"
+#include "structures/gltf_node.h"
 
 class GLTFDocumentExtension : public Resource {
 	GDCLASS(GLTFDocumentExtension, Resource);
