+++ modules/gltf/structures/gltf_camera.cpp
