+++ modules/gltf/structures/gltf_node.cpp
