+++ modules/gltf/structures/gltf_texture.cpp
