+++ modules/gltf/structures/gltf_texture.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  gltf_accessor.h                                                      */
+/*  gltf_texture.h                                                       */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,77 +28,24 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef GLTF_ACCESSOR_H
-#define GLTF_ACCESSOR_H
+#ifndef GLTF_TEXTURE_H
+#define GLTF_TEXTURE_H
 
+#include "../gltf_defines.h"
 #include "core/io/resource.h"
 
-#include "gltf_defines.h"
-
-struct GLTFAccessor : public Resource {
-	GDCLASS(GLTFAccessor, Resource);
-	friend class GLTFDocument;
+class GLTFTexture : public Resource {
+	GDCLASS(GLTFTexture, Resource);
 
 private:
-	GLTFBufferViewIndex buffer_view = 0;
-	int byte_offset = 0;
-	int component_type = 0;
-	bool normalized = false;
-	int count = 0;
-	GLTFType type = GLTFType::TYPE_SCALAR;
-	Vector<double> min;
-	Vector<double> max;
-	int sparse_count = 0;
-	int sparse_indices_buffer_view = 0;
-	int sparse_indices_byte_offset = 0;
-	int sparse_indices_component_type = 0;
-	int sparse_values_buffer_view = 0;
-	int sparse_values_byte_offset = 0;
+	GLTFImageIndex src_image = 0;
 
 protected:
 	static void _bind_methods();
 
 public:
-	GLTFBufferViewIndex get_buffer_view();
-	void set_buffer_view(GLTFBufferViewIndex p_buffer_view);
-
-	int get_byte_offset();
-	void set_byte_offset(int p_byte_offset);
-
-	int get_component_type();
-	void set_component_type(int p_component_type);
-
-	bool get_normalized();
-	void set_normalized(bool p_normalized);
-
-	int get_count();
-	void set_count(int p_count);
-
-	int get_type();
-	void set_type(int p_type);
-
-	Vector<double> get_min();
-	void set_min(Vector<double> p_min);
-
-	Vector<double> get_max();
-	void set_max(Vector<double> p_max);
-
-	int get_sparse_count();
-	void set_sparse_count(int p_sparse_count);
-
-	int get_sparse_indices_buffer_view();
-	void set_sparse_indices_buffer_view(int p_sparse_indices_buffer_view);
-
-	int get_sparse_indices_byte_offset();
-	void set_sparse_indices_byte_offset(int p_sparse_indices_byte_offset);
-
-	int get_sparse_indices_component_type();
-	void set_sparse_indices_component_type(int p_sparse_indices_component_type);
-
-	int get_sparse_values_buffer_view();
-	void set_sparse_values_buffer_view(int p_sparse_values_buffer_view);
-
-	int get_sparse_values_byte_offset();
-	void set_sparse_values_byte_offset(int p_sparse_values_byte_offset);
+	GLTFImageIndex get_src_image() const;
+	void set_src_image(GLTFImageIndex val);
 };
-#endif // GLTF_ACCESSOR_H
+
+#endif // GLTF_TEXTURE_H
